a parte do codigo agora esta refaturada

+++ frontend/script/code.js
@@ -1,27 +1,36 @@
 
+class codigoConfirmacao{
+  static #input = document.querySelector("input").value
+  static #button = document.querySelector("button")
 
-const input = document.querySelector("input")
-const butom = document.querySelector("button")
-
-butom.addEventListener("click", () => {
-  (async () => {
+  static async #confirmarCodigo(){
     try {
-      const httpRequest = await fetch("http://localhost:8080/validar", {
+      const data = await fetch("http://localhost:8080/validar", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({codigo: input.value})
-      });
-      const response = await httpRequest.json()
-      const {jwt} = response
+        body: JSON.stringify({codigo: this.#input})
+      })
+      const {jwt, register} = await data.json()
       localStorage.setItem("jwt", jwt)
-      response.register ? location.href = "/sistema-pagamento/frontend/index.html" : null
+      this.#redirecionarPagina(register)
     } catch (error) {
-        console.log(error);
-        alert('ocorreu um erro inesperado http from.')
+      console.log(error);
     }
-  })();
-});
+  }
 
+  static #redirecionarPagina(response){
+    if (response) {
+      location.href = "/sistema-pagamento/frontend/index.html"
+    }
+  }
 
+  static buttonEvento(){
+    this.#button.addEventListener("click", () =>{
+      codigoConfirmacao.#confirmarCodigo()
+    })
+  }
+}
+
+codigoConfirmacao.buttonEvento()
